# PERF: Speeds up creation of Period, PeriodArray, with Offset freq