@jorisvandenbossche 2018-11-09
This already improves the slowdown (using the example code of the plotting benchmark), like from 5.5 s to 2 s, but it is still a lot slower than on 0.23 (100 ms)

Issues related to previous time this happened: https://github.com/pandas-dev/pandas/issues/11831, https://github.com/pandas-dev/pandas/pull/12909, https://github.com/pandas-dev/pandas/issues/12903)