@jorisvandenbossche 2018-11-09
To be clear, although this PR improved performance a bit, the perf regression is not yet fixed