@jbrockmendel 2018-11-09
> but it is still a lot slower than on 0.23

IIRC the version of `to_offset` in tslibs.offsets was implemented largely for aesthetic purposes, to have the runtime import in just one place.  If there is a perf hit, let's absolutely revert it.

The long-term solution may just be to move `to_offset` up to cython.  The PITA is that requires moving most of tseries.offsets up to cython.