@jorisvandenbossche 2018-11-09
Not saying that the remaining performance hit is necessarily due to `to_offset` (didn't look enough in detail for that), just that there are still performance problems in general with Periods. 

For the plotting, it seems to be coming from Period.asfreq, which uses `to_offset` amongst other things